requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.31
mysql-connector-python==8.4.0
pydantic==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
